pyproject.toml
[tool.poetry]
name = "hitide-backfill-lambdas"
version = "0.0.0a0"
description = ""
authors = ["podaac-tva <podaac-tva@jpl.nasa.gov>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.nasa.gov/podaac/hitide-backfill-lambdas"
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]
exclude = ['contrib', 'docs', 'tests']
packages = [
    { include = "podaac" },
]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^2.0.23"
python-dateutil = "^2.8.2"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^8.0.2"
pylint = "^3.1.0"
flake8 = "^7.0.0"
boto3 = "^1.34.51"
coverage = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-dotenv = "^0.5.2"
moto = "^4.2.10"
PyYAML = "^6.0"
mock = "^5.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
